+++ src/wasm/openjtalk-web/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bumped `@piper-plus/g2p` peer dependency to `^0.3.0`
+- Bumped `@piper-plus/g2p` dependency to `^0.3.0`
 
 ## [0.3.0] - 2026-04-11